requirements.txt
pypdf
transformers
torch
faiss-cpu
fastapi
uvicorn

# Optional providers
openai
google-genai